test: should throw error
