+++ src/tests/utils/cdn-assertions-tests.utils.ts
@@ -1,4 +1,4 @@
-import type { _SERVICE as ConsoleActor } from '$declarations/console/console.did';
+import type { CommitProposal, _SERVICE as ConsoleActor } from '$declarations/console/console.did';
 import type { _SERVICE as MissionControlActor } from '$declarations/mission_control/mission_control.did';
 import type { StorageConfig } from '$declarations/satellite/satellite.did';
 import { toNullable } from '@dfinity/utils';
@@ -33,6 +33,37 @@ export const testNotAllowedCdnMethods = ({
 		const { get_storage_config } = actor();
 
 		await expect(get_storage_config()).rejects.toThrow(errorMsg);
+	});
+
+	it('should throw errors on delete proposal assets', async () => {
+		const { delete_proposal_assets } = actor();
+
+		await expect(delete_proposal_assets({ proposal_ids: [1n] })).rejects.toThrow(errorMsg);
+	});
+
+	it('should throw errors on init proposal', async () => {
+		const { init_proposal } = actor();
+
+		await expect(init_proposal({ AssetsUpgrade: { clear_existing_assets: [] } })).rejects.toThrow(
+			errorMsg
+		);
+	});
+
+	it('should throw errors on submit proposal', async () => {
+		const { submit_proposal } = actor();
+
+		await expect(submit_proposal(123n)).rejects.toThrow(errorMsg);
+	});
+
+	it('should throw errors on commit proposal', async () => {
+		const { commit_proposal } = actor();
+
+		const commit: CommitProposal = {
+			sha256: [1, 2, 3],
+			proposal_id: 123n
+		};
+
+		await expect(commit_proposal(commit)).rejects.toThrow(errorMsg);
 	});
 };
